Splitted ApiLayerTests
