+++ PursuitPal/PusuitPal.Architecture.Tests/TypesResidenceTests.cs
@@ -1,7 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using PursuitPal.Core.Contracts;
-using PursuitPal.Core.Contracts.Repositories;
 using PursuitPal.Infrastructure.Entities;
 using PursuitPal.Presentation.Api.Validators;
 using PusuitPal.Architecture.Tests.Extensions;
@@ -10,9 +8,9 @@ using Xunit;
 namespace PusuitPal.Architecture.Tests
 {
     [Collection("Project Types")]
-    public class ApiLayerTests : BaseTests
+    public class TypesResidenceTests : BaseTests
     {
-        public ApiLayerTests(ProjectTypesFixture fixtures)
+        public TypesResidenceTests(ProjectTypesFixture fixtures)
             : base(fixtures)
         {
         }
@@ -22,22 +20,6 @@ namespace PusuitPal.Architecture.Tests
         {
             AllTypes.That().Inherit(typeof(Controller))
                 .Should().ResideInNamespaceEndingWith("Presentation.Api.Controllers")
-                .AssertIsSuccessful();
-        }
-
-        [Fact]
-        public void AllApiControllersMustInheritController()
-        {
-            AllTypes.That().ResideInNamespaceEndingWith("Presentation.Api.Controllers")
-                .Should().Inherit(typeof(Controller))
-                .AssertIsSuccessful();
-        }
-
-        [Fact]
-        public void ApiControllersShouldEndWithController()
-        {
-            AllTypes.That().Inherit(typeof(Controller))
-                .Should().HaveNameEndingWith("Controller")
                 .AssertIsSuccessful();
         }
 
@@ -58,14 +40,6 @@ namespace PusuitPal.Architecture.Tests
         }
 
         [Fact]
-        public void RequestsShouldEndWithRequest()
-        {
-            AllTypes.That().ResideInNamespaceContaining("Core.Requests")
-                .Should().HaveNameEndingWith("Request")
-                .AssertIsSuccessful();
-        }
-
-        [Fact]
         public void RequestsShouldResideInCore()
         {
             AllTypes.That().HaveNameEndingWith("Request")
@@ -74,26 +48,10 @@ namespace PusuitPal.Architecture.Tests
         }
 
         [Fact]
-        public void ResponsesShouldEndWithResponse()
-        {
-            AllTypes.That().ResideInNamespaceContaining("Core.Responses")
-                .Should().HaveNameEndingWith("Response")
-                .AssertIsSuccessful();
-        }
-
-        [Fact]
         public void ResponsesShouldResideInCore()
         {
             AllTypes.That().HaveNameEndingWith("Response")
                 .Should().ResideInNamespaceContaining("Core.Responses")
-                .AssertIsSuccessful();
-        }
-
-        [Fact]
-        public void ServicesShouldEndWithService()
-        {
-            AllTypes.That().ResideInNamespaceMatching("PursuitPal.Core.Contracts.Services")
-                .Should().HaveNameEndingWith("Service")
                 .AssertIsSuccessful();
         }
 
@@ -110,14 +68,6 @@ namespace PusuitPal.Architecture.Tests
         {
             AllTypes.That().Inherit(typeof(Auditable))
                 .Should().ResideInNamespaceContaining("Infrastructure.Entities")
-                .AssertIsSuccessful();
-        }
-
-        [Fact]
-        public void RepositoriesShouldResideInRepositories()
-        {
-            AllTypes.That().ImplementInterface(typeof(IRepository<>))
-                .Should().ResideInNamespaceContaining("Infrastructure.Repositories")
                 .AssertIsSuccessful();
         }
